Add quaternions + small additions to vector types
Adds the initial implementation of the generic quaternion type (`Quat`),
including common operations such as addition, subtraction,
multiplication, scalar division, magnitude, conjugate, inverse, and
normalization, as well as a axis-angle constructor and function to apply
the quat as a rotation to a `Vec3`.

This commit also adds a few convenience features to the vector types,
including operator assignment with scalars, and additional methods to
convert between vector types, tuples and (sized) arrays.